+++ games/newtonian/game.go
@@ -83,7 +83,7 @@ type Game interface {
 
 	// TimeAddedPerTurn is the amount of time (in nano-seconds) added
 	// after each player performs a turn.
-	TimeAddedPerTurn() int64
+	TimeAddedPerTurn() float64
 
 	// TimeImmune is the number turns a unit is immune to being
 	// stunned.
